+++ packages/radio/src/context.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import type { ReturnContext } from './types';
+
+export const RadioContext = createContext<ReturnContext | null>(null);
